Revert "Move the rush-lib bundle to lib/index.js."
This reverts commit 5a032294132ba6e05b80e87356a80aaba0a7be08.
