--- common/changes/@microsoft/rush/bundle-in-lib_2022-12-12-15-19.json
@@ -1,11 +1,0 @@
-{
-  "changes": [
-    {
-      "comment": "",
-      "type": "none",
-      "packageName": "@microsoft/rush"
-    }
-  ],
-  "packageName": "@microsoft/rush",
-  "email": "iclanton@users.noreply.github.com"
-}
