+++ libraries/rush-lib/config/heft.json
@@ -39,7 +39,7 @@
            * to this folder.
            * NOTE: Assigning "sourceFolder" does not by itself select any files to be copied.
            */
-          "sourceFolder": "lib-commonjs/cli/test/rush-mock-flush-telemetry-plugin",
+          "sourceFolder": "lib/cli/test/rush-mock-flush-telemetry-plugin",
 
           /**
            * (Required) One or more folders that files will be copied into, relative to the project root.
@@ -47,7 +47,7 @@
            * streams to efficiently write multiple outputs.
            */
           "destinationFolders": [
-            "lib-commonjs/cli/test/tapFlushTelemetryAndRunBuildActionRepo/common/autoinstallers/plugins/node_modules/rush-mock-flush-telemetry-plugin"
+            "lib/cli/test/tapFlushTelemetryAndRunBuildActionRepo/common/autoinstallers/plugins/node_modules/rush-mock-flush-telemetry-plugin"
           ],
 
           /**
@@ -63,18 +63,6 @@
            * modifies the link.)  The default value is false.
            */
           "hardlink": true
-        }
-      ]
-    },
-    {
-      "actionKind": "copyFiles",
-      "heftEvent": "bundle",
-      "actionId": "copy-start",
-      "copyOperations": [
-        {
-          "sourceFolder": "lib-commonjs",
-          "destinationFolders": ["lib"],
-          "includeGlobs": ["start.js"]
         }
       ]
     }
