+++ libraries/rush-lib/package.json
@@ -37,8 +37,8 @@
     "colors": "~1.2.1",
     "figures": "3.0.0",
     "git-repo-info": "~2.1.0",
-    "glob": "~7.0.5",
     "glob-escape": "~0.0.2",
+    "glob": "~7.0.5",
     "https-proxy-agent": "~5.0.0",
     "ignore": "~5.1.6",
     "inquirer": "~7.3.3",
@@ -62,9 +62,9 @@
   "devDependencies": {
     "@pnpm/logger": "4.0.0",
     "@rushstack/eslint-config": "workspace:*",
-    "@rushstack/heft": "workspace:*",
     "@rushstack/heft-node-rig": "workspace:*",
     "@rushstack/heft-webpack5-plugin": "workspace:*",
+    "@rushstack/heft": "workspace:*",
     "@rushstack/webpack-preserve-dynamic-require-plugin": "workspace:*",
     "@types/cli-table": "0.3.0",
     "@types/glob": "7.1.1",
