+++ libraries/rush-lib/src/cli/test/Cli.test.ts
@@ -8,7 +8,7 @@ import { Utilities } from '../../utilities/Utilities';
 describe('CLI', () => {
   it('should not fail when there is no rush.json', () => {
     const workingDir: string = '/';
-    const startPath: string = path.resolve(__dirname, '../../../lib-commonjs/start.js');
+    const startPath: string = path.resolve(__dirname, '../../../lib/start.js');
 
     expect(() => {
       Utilities.executeCommand({
@@ -22,7 +22,7 @@ describe('CLI', () => {
 
   it('rushx should pass args to scripts', () => {
     // Invoke "rushx"
-    const startPath: string = path.resolve(__dirname, '../../../lib-commonjs/startx.js');
+    const startPath: string = path.resolve(__dirname, '../../../lib/startx.js');
 
     // Run "rushx show-args 1 2 -x" in the "repo/rushx-project" folder
     const output: string = Utilities.executeCommandAndCaptureOutput(
@@ -40,7 +40,7 @@ describe('CLI', () => {
 
   it('rushx should fail in un-rush project', () => {
     // Invoke "rushx"
-    const startPath: string = path.resolve(__dirname, '../../../lib-commonjs/startx.js');
+    const startPath: string = path.resolve(__dirname, '../../../lib/startx.js');
 
     const output = Utilities.executeCommandAndCaptureOutput(
       'node',
