+++ libraries/rush-lib/src/start.ts
@@ -1,6 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license.
 // See LICENSE in the project root for license information.
 
-import { Rush } from './index';
+import { Rush } from './api/Rush';
 
 Rush.launch(Rush.version, { isManaged: false });
