+++ libraries/rush-lib/tsconfig.json
@@ -3,7 +3,6 @@
   "compilerOptions": {
     "types": ["heft-jest", "node", "webpack-env"],
     "skipLibCheck": true,
-    "resolveJsonModule": true,
-    "outDir": "lib-commonjs"
+    "resolveJsonModule": true
   }
 }
